Move node toolbars into general toolbar directory
