+++ apps/src/sketchlab/reactFlow/nodes/ImageNode.tsx
@@ -5,10 +5,10 @@ import React, {memo, useCallback, useEffect, useRef, useState} from 'react';
 
 import {MIN_NODE_HEIGHT, MIN_NODE_WIDTH} from '../constants';
 import {useSketchLabReadOnly} from '../context';
+import ImageNodeToolbar from '../toolbars/ImageNodeToolbar';
 import {ImageNodeType} from '../types';
 
 import ConnectionHandles from './ConnectionHandles';
-import ImageNodeToolbar from './nodeToolbars/ImageNodeToolbar';
 
 import styles from './image-node.module.scss';
