+++ apps/src/sketchlab/reactFlow/toolbars/LineEdgeToolbar.tsx
@@ -2,12 +2,9 @@ import React from 'react';
 
 import {SketchlabReactFlowEdge} from '@cdo/apps/lab2/types';
 
-import NodeToolbarShell from '../nodes/nodeToolbars/NodeToolbarShell';
-import SwatchGroup from '../nodes/nodeToolbars/SwatchGroup';
-import {
-  DEFAULT_STROKE_COLOR,
-  STROKE_FONT_PALETTE,
-} from '../nodes/nodeToolbars/toolbarPalettes';
+import NodeToolbarShell from './NodeToolbarShell';
+import SwatchGroup from './SwatchGroup';
+import {DEFAULT_STROKE_COLOR, STROKE_FONT_PALETTE} from './toolbarPalettes';
 
 interface LineEdgeToolbarProps {
   edge: SketchlabReactFlowEdge;
