+++ apps/src/sketchlab/reactFlow/toolbars/ShapeNodeToolbar.tsx
@@ -1,35 +1,50 @@
 import React from 'react';
 
-import {TextNodeType} from '../../types';
+import {ShapeNodeType} from '../types';
 
 import FontSizeGroup from './FontSizeGroup';
 import HandleVisibilityToggle from './HandleVisibilityToggle';
 import NodeToolbarShell from './NodeToolbarShell';
 import SwatchGroup from './SwatchGroup';
 import {
+  BACKGROUND_PALETTE,
+  DEFAULT_BACKGROUND_COLOR,
   DEFAULT_FONT_COLOR,
   DEFAULT_FONT_SIZE,
+  DEFAULT_STROKE_COLOR,
   FontSizeValue,
   STROKE_FONT_PALETTE,
 } from './toolbarPalettes';
 import {useNodeToolbarData} from './useNodeToolbarData';
 
-interface TextNodeToolbarProps {
+interface ShapeNodeToolbarProps {
   nodeId: string;
 }
 
-export default function TextNodeToolbar({nodeId}: TextNodeToolbarProps) {
-  const {data, patchNodeData} = useNodeToolbarData<TextNodeType>(nodeId);
+export default function ShapeNodeToolbar({nodeId}: ShapeNodeToolbarProps) {
+  const {data, patchNodeData} = useNodeToolbarData<ShapeNodeType>(nodeId);
 
-  const {fontSize, fontColor} = data;
+  const {backgroundColor, strokeColor, fontSize, fontColor} = data;
   const handlesVisible = data.showHandles !== false;
 
   return (
     <NodeToolbarShell
       target={{type: 'node', id: nodeId}}
       anchorNodeId={nodeId}
-      ariaLabel="Text style"
+      ariaLabel="Shape style"
     >
+      <SwatchGroup
+        groupLabel="Background"
+        swatches={BACKGROUND_PALETTE}
+        selectedValue={backgroundColor ?? DEFAULT_BACKGROUND_COLOR}
+        onSelect={value => patchNodeData({backgroundColor: value})}
+      />
+      <SwatchGroup
+        groupLabel="Border"
+        swatches={STROKE_FONT_PALETTE}
+        selectedValue={strokeColor ?? DEFAULT_STROKE_COLOR}
+        onSelect={value => patchNodeData({strokeColor: value})}
+      />
       <FontSizeGroup
         selectedValue={fontSize ?? DEFAULT_FONT_SIZE}
         onSelect={value => patchNodeData({fontSize: value as FontSizeValue})}
